move `src/template/` to `src/codemods/template/`
