+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/automatic-var-injection.test.ts
