+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/handle-empty-arbitrary-values.ts
@@ -1,5 +1,5 @@
-import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
 
 export function handleEmptyArbitraryValues(
   designSystem: DesignSystem,
