+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/important.ts
@@ -1,6 +1,6 @@
-import { parseCandidate } from '../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { parseCandidate } from '../../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
 import { printCandidate } from '../candidates'
 import { isSafeMigration } from '../is-safe-migration'
