+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/legacy-arbitrary-values.ts
@@ -1,7 +1,7 @@
-import { parseCandidate } from '../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
-import { segment } from '../../../../tailwindcss/src/utils/segment'
+import { parseCandidate } from '../../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
+import { segment } from '../../../../../tailwindcss/src/utils/segment'
 import { printCandidate } from '../candidates'
 
 export function legacyArbitraryValues(
