+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/legacy-classes.ts
@@ -1,10 +1,10 @@
 import { __unstable__loadDesignSystem } from '@tailwindcss/node'
 import path from 'node:path'
 import url from 'node:url'
-import type { Candidate } from '../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
-import { DefaultMap } from '../../../../tailwindcss/src/utils/default-map'
+import type { Candidate } from '../../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
+import { DefaultMap } from '../../../../../tailwindcss/src/utils/default-map'
 import { printCandidate } from '../candidates'
 import { isSafeMigration } from '../is-safe-migration'
