+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/simple-legacy-classes.test.ts
