+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/theme-to-var.test.ts
