+++ packages/@tailwindcss-upgrade/src/codemods/template/codemods/variant-order.ts
@@ -1,7 +1,7 @@
-import { walk, type AstNode } from '../../../../tailwindcss/src/ast'
-import { type Variant } from '../../../../tailwindcss/src/candidate'
-import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { walk, type AstNode } from '../../../../../tailwindcss/src/ast'
+import { type Variant } from '../../../../../tailwindcss/src/candidate'
+import type { Config } from '../../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../../tailwindcss/src/design-system'
 import { printCandidate } from '../candidates'
 
 export function variantOrder(
