+++ packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
