+++ packages/@tailwindcss-upgrade/src/codemods/template/prepare-config.ts
@@ -2,11 +2,11 @@ import { __unstable__loadDesignSystem, compile } from '@tailwindcss/node'
 import fs from 'node:fs/promises'
 import path, { dirname } from 'node:path'
 import { fileURLToPath } from 'node:url'
-import { loadModule } from '../../../@tailwindcss-node/src/compile'
-import { resolveConfig } from '../../../tailwindcss/src/compat/config/resolve-config'
-import type { Config } from '../../../tailwindcss/src/compat/plugin-api'
-import type { DesignSystem } from '../../../tailwindcss/src/design-system'
-import { error, highlight, relative } from '../utils/renderer'
+import { loadModule } from '../../../../@tailwindcss-node/src/compile'
+import { resolveConfig } from '../../../../tailwindcss/src/compat/config/resolve-config'
+import type { Config } from '../../../../tailwindcss/src/compat/plugin-api'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { error, highlight, relative } from '../../utils/renderer'
 import { migratePrefix } from './codemods/prefix'
 
 const __filename = fileURLToPath(import.meta.url)
